+++ cmd/kube-aggregator/pkg/cmd/server/BUILD
@@ -22,13 +22,13 @@ go_library(
         "//pkg/genericapiserver/server:go_default_library",
         "//pkg/genericapiserver/server/options:go_default_library",
         "//pkg/kubectl/cmd/util:go_default_library",
-        "//pkg/storage/storagebackend:go_default_library",
         "//vendor:github.com/pborman/uuid",
         "//vendor:github.com/spf13/cobra",
         "//vendor:k8s.io/apimachinery/pkg/runtime/schema",
         "//vendor:k8s.io/apimachinery/pkg/util/sets",
         "//vendor:k8s.io/apimachinery/pkg/util/wait",
         "//vendor:k8s.io/apiserver/pkg/server/filters",
+        "//vendor:k8s.io/apiserver/pkg/storage/storagebackend",
         "//vendor:k8s.io/client-go/rest",
     ],
 )
