+++ cmd/kube-aggregator/pkg/cmd/server/start.go
@@ -28,6 +28,7 @@ import (
 	"k8s.io/apimachinery/pkg/util/sets"
 	"k8s.io/apimachinery/pkg/util/wait"
 	"k8s.io/apiserver/pkg/server/filters"
+	"k8s.io/apiserver/pkg/storage/storagebackend"
 	restclient "k8s.io/client-go/rest"
 	"k8s.io/kubernetes/cmd/kube-aggregator/pkg/apiserver"
 	"k8s.io/kubernetes/cmd/kube-aggregator/pkg/legacy"
@@ -38,7 +39,6 @@ import (
 	genericapiserver "k8s.io/kubernetes/pkg/genericapiserver/server"
 	genericoptions "k8s.io/kubernetes/pkg/genericapiserver/server/options"
 	cmdutil "k8s.io/kubernetes/pkg/kubectl/cmd/util"
-	"k8s.io/kubernetes/pkg/storage/storagebackend"
 
 	"k8s.io/kubernetes/cmd/kube-aggregator/pkg/apis/apiregistration/v1alpha1"
 )
