+++ examples/apiserver/BUILD
@@ -21,11 +21,11 @@ go_library(
         "//pkg/genericapiserver/server:go_default_library",
         "//pkg/genericapiserver/server/options:go_default_library",
         "//pkg/kubeapiserver/options:go_default_library",
-        "//pkg/storage/storagebackend:go_default_library",
         "//vendor:github.com/golang/glog",
         "//vendor:k8s.io/apimachinery/pkg/runtime/schema",
         "//vendor:k8s.io/apimachinery/pkg/util/errors",
         "//vendor:k8s.io/apiserver/pkg/authorization/authorizerfactory",
+        "//vendor:k8s.io/apiserver/pkg/storage/storagebackend",
     ],
 )
