+++ pkg/controller/serviceaccount/BUILD
@@ -30,7 +30,6 @@ go_library(
         "//pkg/registry/core/serviceaccount:go_default_library",
         "//pkg/registry/core/serviceaccount/storage:go_default_library",
         "//pkg/serviceaccount:go_default_library",
-        "//pkg/storage/storagebackend:go_default_library",
         "//pkg/util/metrics:go_default_library",
         "//pkg/util/workqueue:go_default_library",
         "//vendor:github.com/golang/glog",
@@ -46,6 +45,7 @@ go_library(
         "//vendor:k8s.io/apimachinery/pkg/util/wait",
         "//vendor:k8s.io/apimachinery/pkg/watch",
         "//vendor:k8s.io/apiserver/pkg/endpoints/request",
+        "//vendor:k8s.io/apiserver/pkg/storage/storagebackend",
         "//vendor:k8s.io/client-go/tools/cache",
     ],
 )
