+++ pkg/genericapiserver/server/BUILD
@@ -30,7 +30,6 @@ go_test(
         "//pkg/genericapiserver/registry/rest:go_default_library",
         "//pkg/genericapiserver/server/options:go_default_library",
         "//pkg/storage/etcd/testing:go_default_library",
-        "//pkg/storage/storagebackend:go_default_library",
         "//vendor:github.com/go-openapi/spec",
         "//vendor:github.com/stretchr/testify/assert",
         "//vendor:k8s.io/apimachinery/pkg/api/meta",
@@ -44,6 +43,7 @@ go_test(
         "//vendor:k8s.io/apiserver/pkg/authentication/user",
         "//vendor:k8s.io/apiserver/pkg/authorization/authorizer",
         "//vendor:k8s.io/apiserver/pkg/endpoints/request",
+        "//vendor:k8s.io/apiserver/pkg/storage/storagebackend",
         "//vendor:k8s.io/apiserver/pkg/util/flag",
         "//vendor:k8s.io/client-go/util/cert",
     ],
@@ -74,7 +74,6 @@ go_library(
         "//pkg/genericapiserver/server/mux:go_default_library",
         "//pkg/genericapiserver/server/options:go_default_library",
         "//pkg/genericapiserver/server/routes:go_default_library",
-        "//pkg/storage/storagebackend:go_default_library",
         "//vendor:github.com/coreos/go-systemd/daemon",
         "//vendor:github.com/emicklei/go-restful",
         "//vendor:github.com/emicklei/go-restful/swagger",
@@ -105,6 +104,7 @@ go_library(
         "//vendor:k8s.io/apiserver/pkg/endpoints/request",
         "//vendor:k8s.io/apiserver/pkg/server/filters",
         "//vendor:k8s.io/apiserver/pkg/server/healthz",
+        "//vendor:k8s.io/apiserver/pkg/storage/storagebackend",
         "//vendor:k8s.io/client-go/rest",
         "//vendor:k8s.io/client-go/util/cert",
     ],
