+++ pkg/genericapiserver/server/storage_factory_test.go
@@ -21,10 +21,10 @@ import (
 	"testing"
 
 	"k8s.io/apimachinery/pkg/runtime/schema"
+	"k8s.io/apiserver/pkg/storage/storagebackend"
 	"k8s.io/kubernetes/pkg/api"
 	"k8s.io/kubernetes/pkg/apis/extensions"
 	"k8s.io/kubernetes/pkg/genericapiserver/server/options"
-	"k8s.io/kubernetes/pkg/storage/storagebackend"
 )
 
 func TestUpdateEtcdOverrides(t *testing.T) {
