+++ pkg/kubeapiserver/BUILD
@@ -18,9 +18,9 @@ go_library(
     deps = [
         "//pkg/api:go_default_library",
         "//pkg/genericapiserver/server:go_default_library",
-        "//pkg/storage/storagebackend:go_default_library",
         "//vendor:k8s.io/apimachinery/pkg/runtime",
         "//vendor:k8s.io/apimachinery/pkg/runtime/schema",
+        "//vendor:k8s.io/apiserver/pkg/storage/storagebackend",
         "//vendor:k8s.io/apiserver/pkg/util/flag",
     ],
 )
