+++ pkg/registry/core/rest/BUILD
@@ -14,8 +14,8 @@ go_test(
     library = ":go_default_library",
     tags = ["automanaged"],
     deps = [
-        "//pkg/storage/storagebackend:go_default_library",
         "//vendor:k8s.io/apimachinery/pkg/runtime/schema",
+        "//vendor:k8s.io/apiserver/pkg/storage/storagebackend",
     ],
 )
