+++ pkg/registry/core/service/allocator/storage/BUILD
@@ -18,8 +18,8 @@ go_test(
         "//pkg/registry/core/service/allocator:go_default_library",
         "//pkg/registry/registrytest:go_default_library",
         "//pkg/storage/etcd/testing:go_default_library",
-        "//pkg/storage/storagebackend:go_default_library",
         "//vendor:golang.org/x/net/context",
+        "//vendor:k8s.io/apiserver/pkg/storage/storagebackend",
     ],
 )
 
@@ -34,11 +34,11 @@ go_library(
         "//pkg/registry/core/rangeallocation:go_default_library",
         "//pkg/registry/core/service/allocator:go_default_library",
         "//pkg/storage:go_default_library",
-        "//pkg/storage/storagebackend:go_default_library",
         "//vendor:golang.org/x/net/context",
         "//vendor:k8s.io/apimachinery/pkg/api/errors",
         "//vendor:k8s.io/apimachinery/pkg/runtime",
         "//vendor:k8s.io/apimachinery/pkg/runtime/schema",
+        "//vendor:k8s.io/apiserver/pkg/storage/storagebackend",
     ],
 )
