+++ pkg/registry/registrytest/BUILD
@@ -24,7 +24,6 @@ go_library(
         "//pkg/genericapiserver/registry/rest/resttest:go_default_library",
         "//pkg/storage/etcd:go_default_library",
         "//pkg/storage/etcd/testing:go_default_library",
-        "//pkg/storage/storagebackend:go_default_library",
         "//pkg/storage/testing:go_default_library",
         "//vendor:k8s.io/apimachinery/pkg/api/errors",
         "//vendor:k8s.io/apimachinery/pkg/api/meta",
@@ -35,6 +34,7 @@ go_library(
         "//vendor:k8s.io/apimachinery/pkg/runtime",
         "//vendor:k8s.io/apimachinery/pkg/watch",
         "//vendor:k8s.io/apiserver/pkg/endpoints/request",
+        "//vendor:k8s.io/apiserver/pkg/storage/storagebackend",
     ],
 )
