+++ pkg/storage/etcd/testing/utils.go
@@ -28,9 +28,9 @@ import (
 	"time"
 
 	"k8s.io/apimachinery/pkg/util/wait"
+	"k8s.io/apiserver/pkg/storage/storagebackend"
 	"k8s.io/kubernetes/pkg/storage/etcd/etcdtest"
 	"k8s.io/kubernetes/pkg/storage/etcd/testing/testingcert"
-	"k8s.io/kubernetes/pkg/storage/storagebackend"
 
 	etcd "github.com/coreos/etcd/client"
 	"github.com/coreos/etcd/clientv3"
