--- pkg/storage/storagebackend/BUILD
@@ -1,31 +1,0 @@
-package(default_visibility = ["//visibility:public"])
-
-licenses(["notice"])
-
-load(
-    "@io_bazel_rules_go//go:def.bzl",
-    "go_library",
-)
-
-go_library(
-    name = "go_default_library",
-    srcs = ["config.go"],
-    tags = ["automanaged"],
-    deps = ["//vendor:k8s.io/apimachinery/pkg/runtime"],
-)
-
-filegroup(
-    name = "package-srcs",
-    srcs = glob(["**"]),
-    tags = ["automanaged"],
-    visibility = ["//visibility:private"],
-)
-
-filegroup(
-    name = "all-srcs",
-    srcs = [
-        ":package-srcs",
-        "//pkg/storage/storagebackend/factory:all-srcs",
-    ],
-    tags = ["automanaged"],
-)
