+++ pkg/storage/storagebackend/factory/BUILD
@@ -20,12 +20,12 @@ go_library(
         "//pkg/storage:go_default_library",
         "//pkg/storage/etcd:go_default_library",
         "//pkg/storage/etcd3:go_default_library",
-        "//pkg/storage/storagebackend:go_default_library",
         "//vendor:github.com/coreos/etcd/client",
         "//vendor:github.com/coreos/etcd/clientv3",
         "//vendor:github.com/coreos/etcd/pkg/transport",
         "//vendor:golang.org/x/net/context",
         "//vendor:k8s.io/apimachinery/pkg/util/net",
+        "//vendor:k8s.io/apiserver/pkg/storage/storagebackend",
     ],
 )
 
@@ -38,10 +38,10 @@ go_test(
         "//pkg/api:go_default_library",
         "//pkg/api/testapi:go_default_library",
         "//pkg/storage/etcd/testing/testingcert:go_default_library",
-        "//pkg/storage/storagebackend:go_default_library",
         "//vendor:github.com/coreos/etcd/integration",
         "//vendor:github.com/coreos/etcd/pkg/transport",
         "//vendor:golang.org/x/net/context",
+        "//vendor:k8s.io/apiserver/pkg/storage/storagebackend",
     ],
 )
