+++ pkg/storage/storagebackend/factory/etcd3.go
@@ -17,9 +17,9 @@ limitations under the License.
 package factory
 
 import (
+	"k8s.io/apiserver/pkg/storage/storagebackend"
 	"k8s.io/kubernetes/pkg/storage"
 	"k8s.io/kubernetes/pkg/storage/etcd3"
-	"k8s.io/kubernetes/pkg/storage/storagebackend"
 
 	"github.com/coreos/etcd/clientv3"
 	"github.com/coreos/etcd/pkg/transport"
